ffmpegger: introduce adaptive video resizing
New option: --video_fit which consumes target resolution.
It crops and resizes original video track to fit this resolution.
It is far from optimal, as it is practically impossible to fit
optimally. Anyway it is some kind of combination of scale and crop
video filters and might be improved in future.

It is prohibitted to use --video_fit along with --settings_video option.